+++ src/version2/groups.ts
@@ -110,10 +110,7 @@ export class Groups {
    * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#permissions) required:** *Browse
    * users and groups* [global permission](https://confluence.atlassian.com/x/x4dKLg).
    */
-  async bulkGetGroups<T = Models.PageGroupDetails>(
-    parameters?: Parameters.BulkGetGroups,
-    callback?: never
-  ): Promise<T>;
+  async bulkGetGroups<T = Models.PageGroupDetails>(parameters?: Parameters.BulkGetGroups, callback?: never): Promise<T>;
   async bulkGetGroups<T = Models.PageGroupDetails>(
     parameters?: Parameters.BulkGetGroups,
     callback?: Callback<T>,