+++ src/version2/models/customFieldContextDefaultValueMultipleOption.ts
@@ -1,4 +1,4 @@
-/** Default value for a multi-select custom field. */
+/** The default value for a multi-select custom field. */
 export interface CustomFieldContextDefaultValueMultipleOption {
   /** The ID of the context. */
   contextId: string;