+++ src/version2/models/customFieldValueUpdateDetails.ts
@@ -1,7 +1,7 @@
 import { CustomFieldValueUpdate } from './customFieldValueUpdate';
 
-/** @deprecated Details of updates for a custom field. */
-export interface CustomFieldValueUpdateRequest {
+/** Details of updates for a custom field. */
+export interface CustomFieldValueUpdateDetails {
   /** The list of custom field update details. */
   updates?: CustomFieldValueUpdate[];
 }