+++ src/version2/parameters/createFieldConfiguration.ts
@@ -1,0 +1,3 @@
+import { FieldConfigurationDetails } from '../models';
+
+export interface CreateFieldConfiguration extends FieldConfigurationDetails {}